fix: 正常に更新されない問題を修正
