+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/notes/impl/db/NoteThreadRecordDAO.kt
@@ -4,7 +4,6 @@ import io.objectbox.Box
 import io.objectbox.BoxStore
 import io.objectbox.kotlin.awaitCallInTx
 import io.objectbox.kotlin.boxFor
-import io.objectbox.kotlin.inValues
 import io.objectbox.kotlin.toFlow
 import io.objectbox.query.QueryBuilder
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -22,10 +21,6 @@ open class NoteThreadRecordDAO @Inject constructor(
         boxStore.boxFor()
     }
 
-    private val noteBox: Box<NoteRecord> by lazy {
-        boxStore.boxFor()
-    }
-
     open suspend fun add(context: ThreadRecord) {
         boxStore.awaitCallInTx {
             val exists = noteThreadContextBox.query().equal(
@@ -39,6 +34,12 @@ open class NoteThreadRecordDAO @Inject constructor(
             }
         }
 
+    }
+
+    open suspend fun update(context: ThreadRecord) {
+        boxStore.awaitCallInTx {
+            noteThreadContextBox.put(context)
+        }
     }
 
     open suspend fun appendBlank(noteId: Note.Id): ThreadRecord {
@@ -63,23 +64,6 @@ open class NoteThreadRecordDAO @Inject constructor(
         }!!
     }
 
-    open suspend fun appendDescendants(threadTarget: Note.Id, appendTargets: List<Note.Id>) {
-        appendBlank(threadTarget)
-        boxStore.awaitCallInTx {
-            val context = requireNotNull(findBy(threadTarget))
-            context.descendants.addAll(findByNotes(appendTargets))
-            noteThreadContextBox.put(context)
-        }
-    }
-
-    open suspend fun appendAncestors(threadTarget: Note.Id, appendTargets: List<Note.Id>) {
-        appendBlank(threadTarget)
-        boxStore.awaitCallInTx {
-            val context = requireNotNull(findBy(threadTarget))
-            context.ancestors.addAll(findByNotes(appendTargets))
-            noteThreadContextBox.put(context)
-        }
-    }
 
     open suspend fun clearRelation(targetNote: Note.Id) {
         boxStore.awaitCallInTx {
@@ -97,15 +81,6 @@ open class NoteThreadRecordDAO @Inject constructor(
             NoteRecord.generateAccountAndNoteId(noteId),
             QueryBuilder.StringOrder.CASE_SENSITIVE
         ).build().findFirst()
-    }
-
-    private fun findByNotes(noteIds: List<Note.Id>): List<NoteRecord> {
-        return noteBox.query().inValues(
-            NoteRecord_.accountIdAndNoteId, noteIds.map {
-                NoteRecord.generateAccountAndNoteId(it)
-            }.toTypedArray(),
-            QueryBuilder.StringOrder.CASE_SENSITIVE
-        ).build().find()
     }
 
     @OptIn(ExperimentalCoroutinesApi::class)
